It currently takes ~four and a half minutes to deploy the koto-rekisteri container to each environment. We can make that faster by fiddling with the load balancer configuration, which by default is very conservative.

Changed settings:
- [New container health check](https://docs.aws.amazon.com/AmazonECS/latest/developerguide/load-balancer-healthcheck.html): from 5 OK responses every 30 seconds to 2 OKs in 5 seconds.
- [New container grace period](https://docs.aws.amazon.com/AmazonECS/latest/developerguide/service_definition_parameters.html#sd-networkconfiguration): how long ECS ignores the health checks of a newly started container (to let slow applications start) changed from 60 seconds (CDK default) to 15 seconds.
- [Old container deregistration delay](https://docs.aws.amazon.com/AmazonECS/latest/developerguide/load-balancer-connection-draining.html): open connections are forcibly closed in 5 seconds instead of the default of 300 seconds.